add unit test for reusing results
Adds a unit test to check that a generated kernel which attempts to access the result from the PyTorch reference model in memory fails the correctness check.

If a generated kernel uses empty_like, the CUDA caching allocator can re-use the physical memory of the previously computed result. All the kernel needs to do is return immediately and it will pass the correctness check.

Note that in order to reproduce this, we need to copy the PyTorch output to the CPU and delete the output object. Then empty_like will fetch the physical memory for the output object.